backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
PyPDF2==3.0.1
pycryptodome==3.19.0
nltk==3.8.1
gunicorn==21.2.0 
cd backend && python3 app.py